% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioanalysis.R
\name{ratioanalysis}
\alias{ratioanalysis}
\title{Creates Ratio Analysis for Utah County}
\usage{
ratioanalysis(pred, real)
}
\arguments{
\item{pred}{numeric predicted values from model}

\item{real}{numeric actual values of homes/apts}
}
\description{
Creates Ratio Analysis for Utah County
}
